Create seperate ids files
