+++ src/MOCK_DB_DATA/library-items/.TEMPLATE/_template.ids.ts
@@ -1,0 +1,3 @@
+export const TemplateIds = {
+  Template: undefined,
+}
