+++ src/MOCK_DB_DATA/library-items/alerts/_alerts.ids.ts
@@ -1,0 +1,4 @@
+export const AlertsIds = {
+  Alert: undefined,
+  AlertDismissible: undefined
+}
