+++ src/MOCK_DB_DATA/library-items/badge/_badge.ids.ts
@@ -1,0 +1,4 @@
+export const BadgeIds = {
+  BadgeCircular: undefined,
+  BadgeRounded: undefined
+}
