+++ src/MOCK_DB_DATA/library-items/buttons/_buttons.ids.ts
@@ -1,0 +1,5 @@
+export const ButtonsIds = {
+  ButtonPrimary: undefined,
+  ButtonSecondary: undefined,
+  ButtonFlat: undefined
+}
