+++ src/MOCK_DB_DATA/library-items/color/_color.ids.ts
@@ -1,0 +1,5 @@
+export const ColorIds = {
+  ColorSystem: undefined,
+  ColorBootstrap: undefined,
+  ColorFonts: undefined
+}
