+++ src/MOCK_DB_DATA/library-items/forms/_forms.ids.ts
@@ -1,0 +1,10 @@
+export const FormsIds = {
+  FormInput: undefined,
+  FormTextarea: undefined,
+  FormValidation: undefined,
+  FormCheckbox: undefined,
+  FormRadio: undefined,
+  FormSelect: undefined,
+  FormSwitch: undefined
+}
+
