+++ src/MOCK_DB_DATA/library-items/icons/_icons.ids.ts
@@ -1,0 +1,4 @@
+export const IconsIds = {
+  IconSet: undefined
+}
+
