+++ src/MOCK_DB_DATA/library-items/modal/_modal.ids.ts
@@ -1,0 +1,3 @@
+export const ModalIds = {
+  Modal: undefined
+}
