+++ src/MOCK_DB_DATA/library-items/table/_table.ids.ts
@@ -1,0 +1,5 @@
+export const TableIds = {
+  Table: undefined,
+  TableSelectableRow: undefined,
+  TableSortable: undefined
+}
