+++ src/MOCK_DB_DATA/library-items/tags/_tags.ids.ts
@@ -1,0 +1,7 @@
+export const TagsIds = {
+  TagCategory: undefined,
+  TagFilterOption: undefined,
+  TagMultipleSelection: undefined,
+  TagSingleSelection: undefined,
+  TagState: undefined
+}
