+++ src/MOCK_DB_DATA/library-items/tooltip/_tooltip.ids.ts
@@ -1,0 +1,3 @@
+export const TooltipIds = {
+  Tooltip: undefined
+}
