+++ src/MOCK_DB_DATA/library-items/typography/_typography.ids.ts
@@ -1,0 +1,4 @@
+export const TypographyIds = {
+  HeadingLevels: undefined,
+  TypographicHierarchy: undefined
+}
